requirements.txt
# Core Dependencies
# Requires Python 3.10 or higher

# Deep Learning
torch>=2.0.0
transformers>=4.35.0
datasets>=2.15.0
accelerate>=0.25.0

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.22.0

# SVG Handling
svgwrite>=1.4.3
svgpathtools>=1.6.1
cairosvg>=2.7.1

# Text Processing
regex>=2023.10.3
spacy>=3.7.0

# Data Processing
pandas>=2.1.0
tqdm>=4.66.0

# Experiment Tracking
wandb>=0.16.0
tensorboard>=2.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Optional: Database (for later)
# psycopg2-binary>=2.9.9
# redis>=5.0.1

